backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
bcrypt
gunicorn
requests
stripe
